requirements.txt
Pillow
PySide6
numpy